added api for getting episodes by id, dao for local storage of podcasts with episodes

+++ app/src/main/java/com/example/podcastapp/data/remote/PodcastApi.kt
@@ -33,6 +33,6 @@ interface PodcastApi {
     @GET(PODCASTS)
     suspend fun getPodcastWitEpisodes(
         @Path("id") id: String,
-        @Query("next_episode_pub_date") nextEpisodePubDate: Int,
+        @Query("next_episode_pub_date") nextEpisodePubDate: Long,
     ): PodcastResponseDto
 }
